Move PlayerList Callback handler to own file
